cut pages edition in the right file : pages --> addpage
