+++ src/App.js
@@ -21,6 +21,7 @@ import Dishes from './pages/Dishes';
 import Meals from './pages/Meals';
 import Diets from './pages/Diets';
 import Pages from './pages/Pages';
+import AddPages from './pages/AddPage';
 import Users from './pages/Users';
 import AdminProfil from './pages/AdminProfil';
 import AdminAuth from './pages/AdminAuth';
@@ -86,6 +87,9 @@ const App = () => {
                   </PrivateRoute>
                   <PrivateRoute exact path='/pages'>
                     <Pages />
+                  </PrivateRoute>
+                  <PrivateRoute exact path='/pages/ajouter'>
+                    <AddPages />
                   </PrivateRoute>
                   <PrivateRoute exact path='/users'>
                     <Users />
